lesson03 task done #3

+++ src/main/java/com/company/Calculator.java
@@ -15,7 +15,7 @@ public class Calculator {
     }
 
     public double division(int a, int b) {
-        double c = (double) a / b;
+        double c = ((double) a) / ((double) b);
         return c;
     }
 }
